Create matchmaking
Add appropriate methods to the Robot and Battle models. Add the algorithm itself. Fix a problem with primary key representation differing in quill and SQL.

+++ conf/db/migration/default/V1__Create_initial_tables.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users
 (
-    id       SERIAL PRIMARY KEY,
+    id       BIGSERIAL PRIMARY KEY,
     created  TIMESTAMP   NOT NULL,
     username VARCHAR(15) NOT NULL,
     password VARCHAR(60) NOT NULL
@@ -9,7 +9,7 @@ CREATE TABLE users
 
 CREATE TABLE published_robots
 (
-    id      SERIAL PRIMARY KEY,
+    id      BIGSERIAL PRIMARY KEY,
     created TIMESTAMP NOT NULL,
     code    TEXT      NOT NULL
 );
@@ -18,10 +18,10 @@ CREATE TABLE published_robots
 
 CREATE TABLE robots
 (
-    id        SERIAL PRIMARY KEY,
+    id        BIGSERIAL PRIMARY KEY,
     created   TIMESTAMP   NOT NULL,
-    user_id   INT         NOT NULL REFERENCES users (id),
-    pr_id     INT REFERENCES published_robots (id),
+    user_id   BIGINT      NOT NULL REFERENCES users (id),
+    pr_id     BIGINT REFERENCES published_robots (id),
     name      VARCHAR(15) NOT NULL,
     dev_code  TEXT        NOT NULL,
     automatch BOOL        NOT NULL,
@@ -33,12 +33,12 @@ CREATE TABLE robots
 
 CREATE TABLE battles
 (
-    id        SERIAL PRIMARY KEY,
+    id        BIGSERIAL PRIMARY KEY,
     created   TIMESTAMP  NOT NULL,
-    r1_id     INT        NOT NULL REFERENCES robots (id),
-    r2_id     INT        NOT NULL REFERENCES robots (id),
-    pr1_id    INT        NOT NULL REFERENCES published_robots (id),
-    pr2_id    INT        NOT NULL REFERENCES published_robots (id),
+    r1_id     BIGINT     NOT NULL REFERENCES robots (id),
+    r2_id     BIGINT     NOT NULL REFERENCES robots (id),
+    pr1_id    BIGINT     NOT NULL REFERENCES published_robots (id),
+    pr2_id    BIGINT     NOT NULL REFERENCES published_robots (id),
     ranked    BOOL       NOT NULL,
     winner    VARCHAR(5) NOT NULL,
 --  If `errored` and r1_won/r2_won, then the other robot errored. Otherwise, both errored.
